Bug #1661600 cpufreq: sprd: fix coverity warning in A12
[root cause  ] coverity warning
[changes     ] fix coverity warning in A12
[side effects] none
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: none
[confirm dependent commit]: none
[board]: pike2/sp9863a/ums512
[test case]: dvfs
[reviewers   ] luyao.wu
[change_type ] feature_bugfix
[tag_product ] common

Fix coverity warning in A12.

+++ drivers/cpufreq/sprd-hwdvfs-cpufreq.c
@@ -243,7 +243,7 @@ static int sprd_dev_pm_opp_table_update(struct sprd_cpufreq_info *info)
 		dev_dbg(info->pdev, "add opp: %luHZ-%luUV\n", freq, volt);
 	}
 
-	rate = clu->temp_max_freq * 1000 + 1;
+	rate = (unsigned long)(clu->temp_max_freq * 1000 + 1);
 	while (!IS_ERR(opp = dev_pm_opp_find_freq_ceil(info->cpu_dev, &rate))) {
 		dev_pm_opp_put(opp);
 		dev_pm_opp_remove(info->cpu_dev, rate);
